Add rstest framework for unit tests  (#32)
This PR introduces rstest as main unit testing framework. 
It's added as a development dependency to the project. 
maplit is added for its hashmap macro.

* Core advantages of using rstest
- can save on some boilerplate code
- easier to add new test cases
- #[once] Fixture to make an object initialise once and be 
shared between all tests

*Disadvatages of using rstest
Complex struct initialisation might not be possible in a test case. 
This can still be a no-issue because we can always write normal rust
tests for anything that can't be done using rstest.