Refactor StorageSync (#55)
This PR refactors code around StorageSync, fixes 
instances where `Pathbuf` is more appropriate type 
to for path handling and changes filename such that 
it adds hostname for identification.

StorageSync is only ever required for local sync cycle. 
On s3 sync currently it checks for top level folder inside 
local data directory to retrieve stream names and respective 
paths, this is not ideal so this is changed so that we only 
go through streams that are returned through 'list_streams' 
in s3 storage. This partially fixes #54 but there should be 
more checks in place when loading streams.

Changes:

- Refactors StorageSync
- Change from DirName to StorageDir for path related stuff
- Use StorageDir in s3 sync
- Only check for files on depth 1 when walking tmp directory
- use hostname for file instead of random string