Use existing schema when querying local data. (#86)
Datafusion needs a valid schema to execute a query. If no associated schema
found for registered table then datafusion tries to infer that schema. If
there are no available listings from which schema can be derived then it fails
and returns an error.

This causes entire query to fail and return error response. This is fixed by
giving it a proper schema to work with that is already part of metadata.
There are however ways that this schema is might not be available
( maybe first event has happened yet), then we simply return doing no
changes to RecordBatch.

Fixes https://github.com/parseablehq/console/issues/14