Use Arrow Schema instead of custom schema type.
Replace custom schema type with arrow schema wrapped in Option. If metadata was able to load a schema then it will be of Some variant, in case there is no schema ( new stream ) then the None variant will be used. This reduced number of times serialization to schema happens and this schema will be directly passed on to future event processing and queries.