Append arrow directly to local data
Currently any event that arrives is directly written to data.parquet. It is
easier to do it this way but this ultimatly slows down log ingestion.
To make log ingestion faster a stream of record-batches would act
as an ideal buffer. Appending arrow record-batches directly to a stream is
faster way to go about it but it has it's own challenges.

- querying realtime would require having these record-batches in memory.
- directly appending record-batches from multiple threads to an append only file
is not supported by Arrow StreamWriter. Locking is used for now
- Re-implementation of stats is required
- Cannot fully guarantee state of localstorage if it failed exactly
at move step of localsync. orphan tmp files can occur although that'd very rare