Config datafusion query with schema (#99)
Instead of pre listing all the valid prefixes before 
query execution and letting datafusion infer the 
schema, it is better to pass the schema we already 
have for a given stream and let datafusion do all 
the heavy lifting. In case stream info does not 
have a schema for a given stream then we return 
early with Error suggesting they need to post 
events to this logstream first.