Append arrow directly to local data (#100)
Currently any event that arrives is directly written to 
data.parquet. It is easier to do it this way but this 
ultimately slows down log ingestion.To make log 
ingestion faster a stream of record-batches would 
act as an ideal buffer. 

Appending arrow record-batches directly to a stream 
is faster way to go about it but it has challenges:

- querying realtime would require having these 
record-batches in memory.
- directly appending record-batches from multiple 
threads to an append only file is not supported by 
Arrow StreamWriter. Locking is used for now.
- Re-implementation of stats is required
- Cannot fully guarantee state of local storage if 
it failed exactly at move step of localsync. orphan
tmp files can occur although that'd be rare