Set max thread priority for local-sync (#104)
Set max thread priority for local-sync.

localsync thread is important and should run at exact 
intervals. To ensure this its priority is set to max. Priority 
setting if fails is ignored for now which means if 
thread_priority can't set it to max then all can be done is
log it. Scheduling behaviour is dependant on the underlying 
OS. It is user's responsiblity to make sure that program 
can set thread/process priority. For example on linux you'd 
have to allow `cpu cpuset` cgroup for current user.

Fixes #103