Use register_table for execute on cache (#128)
Registering table with `register_listing_table` uses the listing feature of the
object store to find the files to be processed. Internally this call
`ListingTable::list_files_for_scan` in datafusion which panics when
table path is empty. This code change is similar to how query works for s3.

Also remove hostname from listing options, so querying files
works properly on S3. This is for cases where one server posted 
data and another server is trying to query the data.