Add demo flag for config
Parseable is demo mode by default. After moving to clap 4.0 it became possible
to add a demo mode flag which will run Parseable with default credential and
storage configuration.

Demo flag is exclusive and cannot be used with other configuration options.
A warning is printed when one is running with `--demo` flag

changes
- unify s3 config to main config as sub argument
- add demo flag
- default arguments only in demo mode