Refactor alerts with support for variety of rules
Alerts are stored on metadata map which is behind and rwlock.
This is fine but having to hold writer lock for alert check is not ideal.
This PR refactors code around alerts implementation.

Each alerts rule should be self contained with its own interior mutable state
if it requires any. Interior mutablity pattern allows for alert to be resolved
with shared reference. An alert can now contain different variant of rules
which can be state or stateless.