Introduce Stats API (#191)
Adds stats implementation based on atomic u64 counters. On each event ingestion counter is updated and on every sync parquet file size is added to compressed counter. On call to metadata api for stats it produces a json serializable `Stats` instance which is copy of metadata stats counter. Stats are synced on every s3 sync cycle and is stored inside parseable.json file.