Change local cache directory structure
This PR introduces changes to local directory and removes tmp directory
All conversion is still carried out in s3 sync cycle. StorageDir is used
as primary class for determining what is the data path for a logstream.

Recordbatch file is signified by extention .arrows.

This PR removes startup sync step as it is not required.