Change local cache directory structure (#202)
This PR introduces changes to local directory and removes tmp directory
All conversion is still carried out in s3 sync cycle. StorageDir is used
as primary class for determining what is the data path for a log stream.
RecordBatch file is signified by extension .arrows.

Also remove startup_sync step as it is not required.