Refactor auth flow  (#425)
Argon2 is CPU intensive and yet highly recommended for 
Password hashing. Before this PR, each API call required a hashing
to validate the basic auth header. 

With this PR, we keep a copy of password in memory (only), lazily 
loaded as users make auth call. This allows much faster authentication.

Other changes in this PR:
- Authentication and Authorisation are merged into single middleware.
- Added AuthMap for modular permission lookup

Part of #250