Test inferred schema for json array (#454)
Arrow infer schema internally uses IndexMap to keep track of inferred schema
while updating it through given json objects. Thus the order in which
insertion happens to this map determines the output order for schema fields
as well.

Since the json object in serde_json uses BtreeMap, the iteration is always
ascending by key. This leads to infer_schema producing schema fields in
ascending order when we infer using only one json object.
In case of json array any new fields encountered are added to the last.
Thus output order of infer schema did not provide good enough guarantee.
This is solved in #450. This commit just adds a test for it.