add live tail support for stream (#516)
This PR adds support for live tailing a stream. With this PR,
server exposes a gRPC server on a configurable port (default 8001)
and uses Arrow Flight as the communication protocol. 

As live tail is requested for a stream, Server sends events from 
that steam, before the local staging and parquet conversion. 
This means, server will use the event record batch to create a
flight stream and send over to the client.

fixes #503