feat: add reverse arrow reader (#548)
This PR adds a new arrow reader which can read arrow file in reverse. This is to be 
used while writing to parquet file such that it yields file sorted in timestamp descending. 
Descending timestamp is the natural format for log data where users are interested in 
looking at the latest events first.

This change helps the client tools (console / pb / others) in the sense that they don't 
have to sort this on client side, neither they need to issue queries with `order by` 
keywords.

This PR also add extra time interval to object sync (5s) to ensure there is no race with 
the local_sync function.