Introduce custom table catalog format
This PR introduces a new table format that keeps track of data files in the data storage.
The format is inspired by Apache Iceberg so has similar naming scheme for things.
A snapshot is the main entry point to a table.
A snapshot consists of list of list of url to manifest file and primary time stats for pruning.
A manifest file contains list of all the actual files present along with the file level statistics.
Currently a manifest file is generated per top level partition ( i.e date ).