feat: add custom table catalogue format (#561)
This PR introduces a new table format that keeps track of 
all the data files in the data storage. The format is inspired by 
Apache Iceberg so has similar naming scheme for things.

Snapshot which is stored in the stream metadata file is the 
main entry point to a table. A snapshot essentially is a list of URL 
to manifest file and primary time statistics for pruning said manifest 
during query. A manifest file contains list of all the actual files along 
with their file level statistics. Currently a manifest file is generated 
per top level partition ( i.e date ).

For old data (data ingested by older versions of Parseable) the 
query mechanism falls back to old style of query data.

Signed-off-by: Satyam Singh <trueleo@protonmail.com>
Co-authored-by: Nick <sop.theteachr@gmail.com>