fix: stats enhancement (#799)
Till now the ingested events count and the ingested size would be 
fixed for the lifetime of a stream. Even if older data was deleted by 
retention stats were not updated all, leading to confusion. 

This PR adds a new way to track stats with three different qualifiers for
each stat type. Now we have lifetime, current and deleted qualifiers for
events ingested, ingested size and stored size, so that the users
have much better view of their data and its compression, storage used
etc. This change will be enhanced further with these updated metrics
visible in console. 

Fixes #788