feat: added internal stream "meta" (#801)
This PR adds internal stream "meta" to be used 
in distributed mode only. The query node creates this 
internal stream. It fetches cluster metrics 
from all ingestion nodes, adds the metrics data to an 
event and ingests the event to this internal stream.
This data will be used by console in cluster page to
show the stats per ingestor.