enhancement: add stream type to stream definition (#884)
This PR adds stream_type=UserDefined for user defined streams. 
For internal streams, it adds stream_type=Internal. This is useful
for clients like the Console to decide on how to use the given stream.

As we add other different types of streams, this segregation will 
ensure a clear pattern of usage for different stream types.