refactor: improvements to listing code for readability and less copy (#1025)
These changes don't impact functionality, but they do have performance benefits 
in the form of lowering memory utilization through the removal of a large number 
of clones in places where they are not required, which even in the case of Arc does 
involve a cost. Plus, parts of the code have been reorganized to be functional in 
nature and hence not require frequent iteration over the same members.

As a bonus, comments explaining working have also been included