chore: add null checks (#1355)
A previous PR modified the conditions config to make the value 
field nullable. But the frontend is sometimes sending an empty 
string in case of null. This PR performs both the checks and rejects 
the request in case the operator is is null or is not null and the value 
is something other than null or an empty string.


Co-authored-by: Nikhil Sinha <131262146+nikhilsinhaparseable@users.noreply.github.com>